// src/modules/application/dto/filter/application-sort.dto.ts
export enum ApplicationSort {
    ID = "id",
    
    NAME = "name",
    DESCRIPTION = "description",
    SLUG = "slug",

    STATUS = "status",
    CREATED_BY = "created_by",
    ORGANIZATION_ID = "organization_id",

    TYPE = "type",

    ENVIRONMENT = "environment",

    LOGO_URL = "logo_url",

    HOMEPAGE_URL = "homepage_url",

    METADATA = "metadata",

    RATE_LIMIT = "rate_limit",

    VERSION = "version",
    CREATED_AT = "createdAt",
    UPDATED_AT = "updatedAt",
}